+++ CHANGES.md
@@ -1,3 +1,5 @@
+#### [unreleased]
+* testing admin-notice-dismissal
 
 #### 1.0
 * This is not a real version number and is just for testing purposes.